Move base actions to auto.actions
The abstract action base classes were in the auto package, not the
auto.actions package. There was no clear justification for this layout,
and it lead to unnecessary imports. This commit moves those classes so
that the entire autonomous action library is self-contained.
