+++ src/main/java/org/team114/lib/auto/actions/NullAction.java
@@ -1,7 +1,5 @@
 package org.team114.lib.auto.actions;
 
-import org.team114.lib.auto.Action;
-
 /**
  * An action that does absolutely nothing.
  */
