+++ src/main/java/org/team114/lib/auto/actions/RunnerAction.java
@@ -1,7 +1,5 @@
 package org.team114.lib.auto.actions;
 
-import org.team114.lib.auto.Action;
-
 /**
  * Wraps a runnable, turning it into an action.&nbsp;This should only be used if it's absolutely needed, because
  * a new action class is more reusable. An example of a case when this action is appropriate is for on-the-fly testing,
