+++ src/main/java/org/team114/lib/auto/actions/package-info.java
@@ -1,5 +1,5 @@
 /**
- * Standard {@linkplain org.team114.lib.auto.Action actions} we reuse from year to year. Any action
+ * Standard {@linkplain org.team114.lib.auto.actions.Action actions} we reuse from year to year. Any action
  * that is added here should be generic (that is, it should be applicable to all conceivable games and robot
  * configurations). Other actions should be placed either in the code for the year they are used, or, if they seem
  * of general utility, in a subpackage of this package.
