+++ src/test/java/org/team114/lib/auto/actions/RunnerActionTest.java
@@ -2,7 +2,6 @@ package org.team114.lib.auto.actions;
 
 import org.junit.*;
 import org.mockito.*;
-import org.team114.lib.auto.Action;
 
 import static org.mockito.Mockito.times;
